rename shape converter - file
